deleted cow.js, started moving to new creature constructor
