+++ creature/behavior/basic.js
@@ -6,7 +6,7 @@ function Creature() {}
 
 Creature.prototype.spawn = function(map){
     this.setPosition(Math.round(Math.random()*map.size), 10, Math.round(Math.random()*map.size));
-}
+};
 
 Creature.prototype.jump = function(x) {
     if (x === undefined) x = 1;
@@ -114,6 +114,111 @@ Creature.prototype.procreate = function() {
 
 Creature.prototype.live = function() {
     this.move(creature.moveRandomly(2), 0, creature.moveRandomly(2), map);
+};
+
+Creature.prototype.getFood = function() {
+    var x = this.position.x - 0.5;
+    var z = this.position.z - 0.5;
+
+    //get surrounding area
+    var around = [];
+    this.map.data.forEach(function(row, rowIndex) {
+        if (rowIndex <= x + 3 && rowIndex >= x - 3) {
+            row.forEach(function(cell, cellIndex) {
+                if (cellIndex <= z + 3 && cellIndex >= z - 3)
+                    around.push({
+                        cell: cell,
+                        x: rowIndex,
+                        z: cellIndex
+                    });
+            })
+        }
+    })
+    //get closest food voxel
+    var min, closestCell;
+    around.forEach(function(cell) {
+        if (cell.cell.material === "grass") {
+            var dist = Math.abs(x + z - (cell.x + cell.z));
+            if (!min) {
+                min = dist;
+                closestCell = cell;
+            } else if (dist < min) {
+                min = dist;
+                closestCell = cell;
+            }
+
+        }
+    })
+
+    this.singleFood = closestCell || "none"
+}
+
+Creature.prototype.moveRandomly = function(amt) {
+    this.moving = true;
+    var x = Math.round(Math.random() * amt) - amt / 2;
+    var z = Math.round(Math.random() * amt) - amt / 2;
+    this.move(x, 0, z)
+}
+
+Creature.prototype.step = function(dir, str) {
+    if (dir < this.singleFood[str]) return 1;
+    else if (dir > this.singleFood[str]) return -1;
+    else return 0;
+};
+
+Creature.prototype.moveTowardsFood = function() {
+    var x = this.position.x - 0.5;
+    var z = this.position.z - 0.5;
+
+    //if moving 0 in all directions
+    if (!this.step(x, "x") && !this.step(z, "z")) {
+        this.foundFood = true;
+        this.moving = false;
+    } else {
+        this.move(this.step(x, "x"), 0, this.step(z, "z"));
+    }
+};
+
+Creature.prototype.findFood = function() {
+    if (this.singleFood === "none") {
+        this.getFood();
+        this.moving = true;
+    }
+
+    if (this.singleFood === "none") {
+        this.moveRandomly(2);
+    }
+    if (typeof this.singleFood === "object" && this.moving) {
+        this.moveTowardsFood();
+    }
+    if (this.foundFood) {
+        if (this.eatingCount === 4) {
+            this.eat();
+            this.hp++
+            this.hpCount = 0;
+            this.getFood();
+            this.eatingCount = 0;
+            this.moving = true;
+        } else this.eatingCount++;
+    }
+};
+
+Creature.prototype.exist = function() {
+    if (this.alive) {
+        this.hpCount++;
+        if (this.hpCount === 10) {
+            this.hp--;
+            this.hpCount = 0;
+        }
+
+        if (this.hp === 0) this.alive = false;
+        if (this.hp <= 5) this.hungry = true;
+        if (this.hp === 10) this.hungry = false;
+
+        if (this.hungry) this.findFood();
+        else this.moveRandomly(2);
+    } else this.alive = false;
+
 };
 
 module.exports = Creature;
